refactor: centralize CLI entrypoint
