+++ src/scaleforge/cli/main.py
@@ -160,6 +160,82 @@ def run_cmd(
     raise SystemExit(0 if ok else 1)
 
 
+# ---------------------------------------------------------------------------
+# Demo commands
+
+
+@cli.command("demo-upscale")
+@click.option("-i", "--input", required=True, type=click.Path(exists=True), help="Input image path")
+@click.option("-o", "--output", required=True, help="Output image path")
+@click.option("-s", "--scale", type=float, default=2.0, show_default=True, help="Upscale factor")
+@click.option(
+    "--mode",
+    type=click.Choice(["nearest", "bilinear", "bicubic", "lanczos"]),
+    default="lanczos",
+    show_default=True,
+    help="Upsampling mode",
+)
+def demo_upscale(input: str, output: str, scale: float, mode: str) -> None:
+    """Demo command for single image upscaling using Pillow"""
+    from scaleforge.demo.upscale import upscale_image
+
+    try:
+        upscale_image(input_path=input, output_path=output, scale=scale, mode=mode)
+    except Exception as e:  # pragma: no cover - runtime error path
+        click.echo(f"Error: {e}", err=True)
+        raise SystemExit(1)
+
+
+@cli.command("demo-batch")
+@click.option("-i", "--input-dir", required=True, type=click.Path(exists=True), help="Input directory")
+@click.option("-o", "--output-dir", required=True, help="Output directory")
+@click.option("-s", "--scale", type=float, default=2.0, show_default=True, help="Upscale factor")
+@click.option(
+    "--mode",
+    type=click.Choice(["nearest", "bilinear", "bicubic", "lanczos"]),
+    default="lanczos",
+    show_default=True,
+    help="Upsampling mode",
+)
+@click.option("--suffix", default="@scaled", show_default=True, help="Filename suffix for outputs")
+@click.option("--include", multiple=True, help="Glob pattern for files to include")
+@click.option("--exclude", multiple=True, help="Glob pattern for files to exclude")
+@click.option("--limit", type=int, help="Maximum number of files to process")
+@click.option("--dry-run", is_flag=True, help="Preview changes without processing")
+@click.option("--overwrite", is_flag=True, help="Overwrite existing output files")
+def demo_batch(
+    input_dir: str,
+    output_dir: str,
+    scale: float,
+    mode: str,
+    suffix: str,
+    include: tuple[str],
+    exclude: tuple[str],
+    limit: int,
+    dry_run: bool,
+    overwrite: bool,
+) -> None:
+    """Batch process images with demo upscaling"""
+    from scaleforge.demo.upscale import batch_upscale
+
+    try:
+        batch_upscale(
+            input_dir=input_dir,
+            output_dir=output_dir,
+            scale=scale,
+            mode=mode,
+            suffix=suffix,
+            include_patterns=include,
+            exclude_patterns=exclude,
+            limit=limit,
+            dry_run=dry_run,
+            overwrite=overwrite,
+        )
+    except Exception as e:  # pragma: no cover - runtime error path
+        click.echo(f"Error: {e}", err=True)
+        raise SystemExit(1)
+
+
 # Global configuration populated during ``cli`` invocation
 _CFG = None
